+++ lib/opam.ml
@@ -1,13 +1,43 @@
 open Import
 
-let depends_on_dune ~allow_jbuilder (formula : OpamTypes.filtered_formula) =
-  let dune = OpamPackage.Name.of_string "dune" in
-  let jbuilder = OpamPackage.Name.of_string "jbuilder" in
-  let is_duneish name =
-    let eq n n' = OpamPackage.Name.compare n n' = 0 in
-    eq dune name || (allow_jbuilder && eq jbuilder name)
+let depends_on_any packages (formula : OpamTypes.filtered_formula) =
+  let packages = List.map ~f:OpamPackage.Name.of_string packages in
+  let is_one_of_packages name =
+    List.exists packages ~f:(fun p -> OpamPackage.Name.compare p name = 0)
   in
-  OpamFormula.fold_left (fun acc (name, _) -> acc || is_duneish name) false formula
+  OpamFormula.fold_left (fun acc (name, _) -> acc || is_one_of_packages name) false formula
+
+let ocaml_options =
+  (* ocaml-options-vanilla is purposefully excluded from that list as it doesn't imply
+     a dependency towards ocaml-variants on its own. *)
+  [
+    "ocaml-option-32bit";
+    "ocaml-option-afl";
+    "ocaml-option-bytecode-only";
+    "ocaml-option-default-unsafe-string";
+    "ocaml-option-flambda";
+    "ocaml-option-fp";
+    "ocaml-option-musl";
+    "ocaml-option-nnp";
+    "ocaml-option-nnpchecker";
+    "ocaml-option-no-flat-float-array";
+    "ocaml-option-spacetime";
+    "ocaml-option-static";
+    "ocaml-options-only-afl";
+    "ocaml-options-only-flambda";
+    "ocaml-options-only-flambda-fp";
+    "ocaml-options-only-fp";
+    "ocaml-options-only-nnp";
+    "ocaml-options-only-nnpchecker";
+    "ocaml-options-only-no-flat-float-array";
+  ]
+
+let depends_on_compiler_variants formula =
+  depends_on_any ("ocaml-variants" :: ocaml_options) formula
+
+let depends_on_dune ~allow_jbuilder (formula : OpamTypes.filtered_formula) =
+  let packages = if allow_jbuilder then [ "dune"; "jbuilder" ] else [ "dune" ] in
+  depends_on_any packages formula
 
 let pull_tree ~url ~hashes ~dir global_state =
   let dir_str = Fpath.to_string dir in