+++ CHANGES.md
@@ -14,6 +14,8 @@
   dune. (#176, @NathanReb)
 - Only print the full list of selected root packages once and only in verbose mode, simply printing
   the number in the default logs. (#173, @NathanReb)
+- Improve the solving process so it only accepts base-compilers unless one explicitly requires
+  an compiler variant, either directly or using `ocaml-option-*` packages. (#178, @NathanReb)
 
 ### Deprecated